pyproject.toml
[build-system]
requires = ["setuptools>=70", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "collector"
version = "0.1.0"
description = "OpenCV-сборщик данных: окно камеры + правая панель атрибутов с кириллицей"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "opencv-python>=4.10",
    "Pillow>=10",
    "rich>=13.0",
    "loguru>=0.7",
    "gradio>=4.0",
    "httpx[socks]",
    "pyyaml>=6.0",
    "pandas>=2.0"  # For CSV export
]

[project.scripts]
collector = "collector.app:main"

[tool.setuptools.packages.find]
where = ["src"]
